requirements.txt
numpy<2.0.0,>=1.19.0
opencv-python>=4.5.0
scipy>=1.8.0
pyyaml>=5.4
torch==1.13.1+cu117
torchvision==0.14.1+cu117
torchaudio==0.13.1
pybind11>=2.9.0
cython>=0.29
matplotlib>=3.5.0
nose>=1.3.0
coverage>=5.3
pytest>=7.0.0
pytest-cov>=2.12.0
black>=20.8b1
flake8>=3.9.0
mypy>=0.900
pre-commit>=2.13.0
opencv-contrib-python>=4.5.0